# server/openapi.yaml
openapi: 3.0.0
info:
  title: Communicator API 2026
  description: >
    API REST + WebSocket para sistema de mensajería instantánea y
    llamadas peer-to-peer (WebRTC). Soporta chat 1-a-1, grupos,
    presencia en línea, y señalización de llamadas de voz/video.
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Servidor Local de Desarrollo (ADB reverse)
  - url: http://192.168.XX.XX:3000
    description: Red WiFi local (reemplazar XX.XX con IP del servidor)
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT obtenido en /api/auth/login o /api/auth/register
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          example: Credenciales inválidas
    ValidationError:
      type: object
      properties:
        errors:
          type: array
          items:
            type: object
            properties:
              msg:
                type: string
              path:
                type: string
              location:
                type: string
...
    User:
      type: object
      required: [id, username, email, avatar_url, is_online, last_seen, tenant_id, active_workspace_id, is_suspended]
      properties:
        id:
          type: integer
          example: 1
        username:
          type: string
          minLength: 3
          maxLength: 50
          pattern: '^[a-zA-Z0-9_]+$'
          example: juan_perez
        email:
          type: string
          format: email
          example: juan@example.com
        avatar_url:
          type: string
          nullable: true
          example: null
        is_online:
          type: boolean
          example: true
        last_seen:
          type: string
          format: date-time
          nullable: true
        tenant_id:
          type: integer
          nullable: true
          description: ID del tenant al que pertenece (alias de compat; hoy = workspace_id)
          example: 1
        active_workspace_id:
          type: integer
          nullable: true
          description: Workspace activo del usuario (seleccionado)
          example: 2
        workspace_ids:
          type: array
          nullable: true
          description: IDs de todos los workspaces a los que pertenece el usuario
          items:
            type: integer
          example: [2, 5]
        is_suspended:
          type: boolean
          description: Si la cuenta está suspendida (login bloqueado)
          example: false
...
    PushTokenRequest:
      type: object
      required: [token, platform]
      properties:
        token:
          type: string
          description: Expo push token (ExponentPushToken[xxxx])
          example: ExponentPushToken[xxxxxxxxxxxxxxxxxxxxxx]
        platform:
          type: string
          enum: [ios, android]
          description: Plataforma del dispositivo
    PushTokenResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
          example: Token registrado exitosamente
...
    Message:
      type: object
      required: [id, sender_id, receiver_id, group_id, content, message_type, created_at, read_at, reply_to_id]
      properties:
        id:
          type: integer
          example: 42
        sender_id:
          type: integer
          example: 1
        receiver_id:
          type: integer
          nullable: true
          example: 2
        group_id:
          type: integer
          nullable: true
          example: null
        content:
          type: string
          example: Hola, ¿cómo estás?
        message_type:
          type: string
          enum: [text, image, system]
          default: text
        created_at:
          type: string
          format: date-time
        read_at:
          type: string
          format: date-time
          nullable: true
        sender_name:
          type: string
          example: juan_perez
        sender_avatar:
          type: string
          nullable: true
        group_name:
          type: string
          nullable: true
          description: Nombre del grupo (solo para mensajes grupales)
        reply_to_id:
          type: integer
          nullable: true
          example: 12
          description: ID del mensaje al que responde
        replied_to:
          type: object
          nullable: true
          required: [id, sender_id, sender_name, content, message_type]
          properties:
            id:
              type: integer
              example: 12
            sender_id:
              type: integer
              example: 1
            sender_name:
              type: string
              example: juan_perez
            content:
              type: string
              example: Mensaje original
            message_type:
              type: string
              enum: [text, image, system]
          description: Mensaje original al que se responde (incluido cuando reply_to_id no es null)
    Conversation:
      type: object
      required: [contact_id, username, avatar_url, is_online, last_message, last_message_at, message_type]
      properties:
        contact_id:
          type: integer
          example: 2
        username:
          type: string
          example: maria_garcia
        avatar_url:
          type: string
          nullable: true
        is_online:
          type: boolean
          example: true
        last_message:
          type: string
          example: Nos vemos mañana
        last_message_at:
          type: string
          format: date-time
        message_type:
          type: string
          enum: [text, image, system]
        unread_count:
          type: integer
          example: 3
    MessagesResponse:
      type: object
      required: [messages, hasMore]
      properties:
        messages:
          type: array
          items:
            $ref: '#/components/schemas/Message'
        hasMore:
          type: boolean
          description: true si hay más páginas disponibles
...
    PushTokenInfo:
      type: object
      properties:
        token:
          type: string
          description: Token truncado (primeros 30 chars)
        platform:
          type: string
          enum: [ios, android]
        isValid:
          type: boolean
          description: Si el token tiene formato Expo válido
    PushTokensResponse:
      type: object
      properties:
        userId:
          type: integer
        count:
          type: integer
        tokens:
          type: array
          items:
            $ref: '#/components/schemas/PushTokenInfo'
    PushTestRequest:
      type: object
      properties:
        userId:
          type: integer
          description: ID del destinatario (opcional, por defecto es el propio usuario)
        title:
          type: string
          description: Título de la notificación
        body:
          type: string
          description: Cuerpo de la notificación
    PushTestResponse:
      type: object
      properties:
        targetUserId:
          type: integer
        tokensFound:
          type: integer
        tokensAreValid:
          type: boolean
        sent:
          type: integer
        failed:
          type: integer
        details:
          type: array
          items:
            type: string
        hint:
          type: string
  responses:
    Unauthorized:
      description: Token JWT requerido o inválido
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
paths:
...
  /api/users/contacts:
    get:
      summary: Listar todos los usuarios
      description: |
        Devuelve todos los usuarios registrados excepto el autenticado,
        ordenados alfabéticamente por username.
      tags: [Usuarios]
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Lista de contactos
          content:
            application/json:
              schema:
                type: object
                properties:
                  contacts:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/users/search:
    get:
      summary: Buscar usuarios
      description: |
        Busca usuarios por username o email.
        - Mínimo 2 caracteres de consulta
        - Máximo 20 resultados
      tags: [Usuarios]
      security:
        - BearerAuth: []
      parameters:
        - name: q
          in: query
          required: true
          description: Término de búsqueda (mín. 2 caracteres)
          schema:
            type: string
            minLength: 2
      responses:
        '200':
          description: Resultados de la búsqueda
          content:
            application/json:
              schema:
                type: object
                properties:
                  results:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
        '400':
          description: Consulta muy corta
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/users/push-token:
    put:
      summary: Registrar token de push notification
      description: |
        Registra o actualiza el Expo push token del dispositivo.
        Si el token ya existe para otro usuario, se reasigna.
      tags: [Usuarios]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PushTokenRequest'
      responses:
        '200':
          description: Token registrado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PushTokenResponse'
        '400':
          description: Error de validación
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ══════════════════════════════════════════════
  # MÓDULO DE MENSAJES
  # ══════════════════════════════════════════════
  /api/messages/conversations:
    get:
      summary: Listar conversaciones
      description: |
        Devuelve todas las conversaciones 1-a-1 del usuario autenticado
        con el último mensaje de cada una, ordenadas por fecha descendente.
      tags: [Mensajes]
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Lista de conversaciones
          content:
            application/json:
              schema:
                type: object
                properties:
                  conversations:
                    type: array
                    items:
                      $ref: '#/components/schemas/Conversation'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/messages/{userId}:
    get:
      summary: Obtener mensajes con un usuario
      description: |
        Mensajes entre el usuario autenticado y otro usuario.
        Paginación por offset (descendente por fecha, entregados en orden ascendente).
      tags: [Mensajes]
      security:
        - BearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
          description: ID del otro usuario
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
          description: Desplazamiento para paginación
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            maximum: 100
          description: Máximo de mensajes por página
      responses:
        '200':
          description: Lista de mensajes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessagesResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ══════════════════════════════════════════════
  # MÓDULO DE GRUPOS
  # ══════════════════════════════════════════════
...
  /api/users/push-tokens:
    get:
      summary: Listar push tokens del usuario (debug)
      description: |
        Devuelve los Expo push tokens registrados para el usuario autenticado.
        Solo uso en depuración para verificar que los tokens se registraron correctamente.
      tags: [Debug]
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Lista de tokens registrados
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PushTokensResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/push/test:
    post:
      summary: Enviar push de prueba (debug)
      description: |
        Envía una notificación push de prueba al usuario especificado
        (o al propio usuario autenticado si no se especifica).
        Solo uso en depuración.
      tags: [Debug]
      security:
        - BearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PushTestRequest'
      responses:
        '200':
          description: Resultado del envío de prueba
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PushTestResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...